Modified logig redirect route
