+++ src/app/api/auth/[auth0]/route.ts
@@ -2,9 +2,6 @@ import { handleAuth, handleLogin } from '@auth0/nextjs-auth0';
 
 export const GET = handleAuth({
   login: handleLogin({
-    authorizationParams: {
-      audience: 'https://deep-pharma.com',
-      scope: 'openid profile email',
-    },
+    returnTo: '/home',
   }),
 });
